requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clearml==1.9.1
click==8.1.3
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
dm-tree==0.1.8
dnspython==2.2.1
easydict==1.13
einops==0.6.1
exceptiongroup==1.0.4
executing==2.0.1
expecttest==0.1.4
face-alignment==1.4.1
fastjsonschema==2.19.1
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.1
finch-clust==0.2.0
fire==0.5.0
flatbuffers==24.3.25
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.5.0
ftfy==6.2.0
furl==2.1.3
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.61.0
idna==3.6
image-dataset-viz==0.2.2
imageio==2.25.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.10.1
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.8
mtcnn-onnxruntime==0.0.1
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
notebook==7.0.7
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-dali-cuda110==1.36.0
# nvidia-nvimgcodec-cu11==0.2.0.7
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.3
onnxruntime-gpu==1.17.1
openai-clip==1.0.1
# opencv-python==4.9.0.80
# opencv-python-headless==4.7.0.68
orderedmultidict==1.0.1
overrides==7.6.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
pooch==1.8.1
portalocker==2.8.2
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-config-runner==0.3.2
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyDeprecate==0.3.2
Pygments==2.17.2
PyJWT==2.4.0
pynndescent==0.5.11
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pytorch-ignite==0.4.11
pytorchvideo==0.1.5
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
submitit==1.5.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.2.0
terminado==0.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.6.13
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
torch-tb-profiler==0.4.3
torch_geometric==2.5.3
torchaudio==0.13.1
torchelastic==0.2.2
torchmetrics==0.8.0
torchsampler==0.1.2
torchstat==0.0.7
torchsummary==1.5.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-dataclasses==0.6.6
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
widgetsnbextension==4.0.9
xlwt==1.3.0
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0